provide type definitions for all exports
