+++ index.d.ts
@@ -1,26 +1,7 @@
-declare namespace filenamify {
-	interface Options {
-		/**
-		String to use as replacement for reserved filename characters.
+import filenamify = require('./filenamify');
+import filenamifyPath = require('./filenamify-path');
 
-		Cannot contain: `<` `>` `:` `"` `/` `\` `|` `?` `*`
-
-		@default '!'
-		*/
-		readonly replacement?: string;
-
-		/**
-		Truncate the filename to the given length.
-		
-		Systems generally allow up to 255 characters, but we default to 100 for usability reasons.
-
-		@default 100
-		*/
-		readonly maxLength?: number;
-	}
-}
-
-declare const filenamify: {
+declare const filenamifyCombined: { 
 	/**
 	Convert a string to a valid filename.
 
@@ -37,10 +18,7 @@ declare const filenamify: {
 	*/
 	(string: string, options?: filenamify.Options): string;
 
-	/**
-	Convert the filename in a path a valid filename and return the augmented path.
-	*/
-	path(path: string, options?: filenamify.Options): string;
+	path: typeof filenamifyPath;
 };
 
-export = filenamify;
+export = filenamifyCombined;
